% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abrome.R
\name{abrome}
\alias{abrome}
\title{Bayesian Region of Measurement Equivalence (ROME) with alignment}
\usage{
abrome(pninv, n_items, n_ninv, data, ROME_lim, eta_lim, eta_num)
}
\arguments{
\item{pninv}{Proportion of noninvariance}

\item{n_items}{Number of items}

\item{n_ninv}{Number of noninvariant items}

\item{data}{The dataset}

\item{ROME_lim}{Preset Region of Measurement Equivalence (ROME)}

\item{eta_lim}{Range of latent ability}

\item{eta_num}{Number of latent ability level}
}
\value{
A vector of length 6.
\item{ROME}{preset ROME}
\item{mean}{posterior mean of the expected group difference on total test scores}
\item{sd}{posterior SD of the expected group difference on total test scores}
\item{PDImax}{maximum length of HPDI for the expected group difference on total test scores}
\item{PDImean}{mean length of HPDI for the expected group difference on total test scores}
\item{decision}{decision on whether the scale is practically invariant across the group of interest}
}
\description{
\code{abrome} implements the Bayesian ROME framework with alignment (ABROME)
}
