% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balign.R
\name{balign}
\alias{balign}
\title{Bayesian alignment}
\usage{
balign(data, n_items, n_ninv, n_gps = 2, svp = FALSE)
}
\arguments{
\item{data}{The dataset}

\item{n_items}{Number of items}

\item{n_ninv}{Number of noninvariant items}

\item{n_gps}{Number of groups}

\item{svp}{Small variance priors}
}
\value{
A vector of length 3.
\item{lambda}{aligned factor loadings estimates}
\item{n_gps}{number of groups}
\item{nu}{aligned intercepts estimates}
}
\description{
\code{balign} implements the Bayesian alignment optimization
}
