% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{exp_diff_cum}
\alias{exp_diff_cum}
\title{Computing the cumulative expected group difference in total test score}
\usage{
exp_diff_cum(eta, si, sl, sil, prob = 0.9)
}
\arguments{
\item{eta}{Latent ability}

\item{si}{Difference in intercepts}

\item{sl}{Difference in loadings}

\item{sil}{The product of the difference in intercepts and difference in loadings}

\item{prob}{the percentage of HPDI}
}
\value{
A vector of length 5.
\item{eta}{latent ability}
\item{mean}{posterior mean total test scores}
\item{sd}{posterior standard deviation of the total test scores}
\item{ll}{lower bound of HPDI}
\item{ul}{upper bound of HPDI}
}
\description{
\code{exp_diff_cum} computes the cumulative expected group difference in
total test score for all groups
}
