% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{exp_tot}
\alias{exp_tot}
\title{Computing the expected total test score}
\usage{
exp_tot(eta, si, sl, prob = 0.95)
}
\arguments{
\item{eta}{Latent ability}

\item{si}{Intercepts for the group of interest}

\item{sl}{Loadings for the group of interest}

\item{prob}{the percentage of HPDI}
}
\value{
A vector of length 5.
\item{eta}{latent ability}
\item{mean}{posterior mean total test scores}
\item{sd}{posterior standard deviation of the total test scores}
\item{ll}{lower bound of HPDI}
\item{ul}{upper bound of HPDI}
}
\description{
\code{exp_tot} computes the expected total test score for each group
}
