% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_expt}
\alias{plot_expt}
\title{Plot the expected total test score for each group}
\usage{
plot_expt(df_post)
}
\arguments{
\item{df_post}{The expected total test scores for the group of interest}
}
\value{
A plot that shows the expected total test scores for the group of interest across latent ability level
}
\description{
\code{plot_expt} plots the expected total test score for each group
}
