% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_exptDiff}
\alias{plot_exptDiff}
\title{Plot the expected difference in total test scores across groups}
\usage{
plot_exptDiff(diff_post, ROME)
}
\arguments{
\item{diff_post}{The expected difference in total test scores across groups}

\item{ROME}{The preset region of measurement equivalence (ROME)}
}
\value{
A plot that shows the expected total test scores for the group of interest across latent ability level
}
\description{
\code{plot_exptDiff} plots the expected difference in total test scores across groups
}
