% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{pooled_sd_test}
\alias{pooled_sd_test}
\title{Computing the pooled standard deviation of total test scores}
\usage{
pooled_sd_test(items, group)
}
\arguments{
\item{items}{Item names for the test.}

\item{group}{The column for the grouping characteristics}
}
\value{
A single number
}
\description{
\code{pooled_sd_test} computes the pooled SD of total test scores
}
